+++ Runtime/Model/Attributes/ProcessAttributeProvider.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Globalization;
 using UnityEngine;
+using UnityEngine.Profiling;
 
 namespace Backtrace.Unity.Model.Attributes
 {
@@ -15,6 +16,13 @@ namespace Backtrace.Unity.Model.Attributes
             }
             attributes["gc.heap.used"] = GC.GetTotalMemory(false).ToString(CultureInfo.InvariantCulture);
             attributes["process.age"] = Math.Round(Time.realtimeSinceStartup).ToString(CultureInfo.InvariantCulture);
+            attributes["system.memory.active"] = Profiler.GetTotalAllocatedMemoryLong().ToString(CultureInfo.InvariantCulture);
+            attributes["system.memory.reserved"] = Profiler.GetTotalReservedMemoryLong().ToString(CultureInfo.InvariantCulture);
+            attributes["system.memory.unused"] = Profiler.GetTotalUnusedReservedMemoryLong().ToString(CultureInfo.InvariantCulture);
+            attributes["system.memory.temp"] = Profiler.GetTempAllocatorSize().ToString(CultureInfo.InvariantCulture);
+            attributes["mono.heap"] = Profiler.GetMonoHeapSizeLong().ToString(CultureInfo.InvariantCulture);
+            attributes["mono.used"] = Profiler.GetMonoUsedSizeLong().ToString(CultureInfo.InvariantCulture);
+
         }
     }
 }